+++ spec/dummy/db/migrate/20150630011645_add_response_code_to_asyncapi_client_jobs.asyncapi_client.rb
@@ -1,0 +1,6 @@
+# This migration comes from asyncapi_client (originally 20150630004215)
+class AddResponseCodeToAsyncapiClientJobs < ActiveRecord::Migration
+  def change
+    add_column :asyncapi_client_jobs, :response_code, :integer
+  end
+end