+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+# TBA
+
+- Add `Asyncapi::Client::Job#response_code`
+
 # 0.2.0
 
 - Mention that `:status` is one of the attributes that should be allowed during mass assignment when using `protected_attributes`